performance-tracker/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.25.0
google-auth>=2.23.0
firebase-admin>=6.2.0
google-cloud-firestore>=2.11.0
pydantic>=2.0.0
python-dotenv>=1.0.0
python-dateutil>=2.8.0